Redefine all references to score as unsigned char
Reduce time for solution after 15 moves from 6.5 to 5.4 and
permits calculation for solution after 14 moves in around
a minute

+++ Memo.cpp
@@ -4,14 +4,14 @@
 Memo::Memo():MAX_SIZE(10'000'000){
 };
 
-void Memo::Add(const Board& b, short score){
+void Memo::Add(const Board& b, uint8_t score){
     mem[b.GetKey()]=score;
     int size = mem.size();
     if(mem.size()>MAX_SIZE){
         mem.erase(mem.begin());
     }
 }
-bool  Memo::TryGet(const Board& b,short &score)const{
+bool  Memo::TryGet(const Board& b,uint8_t &score)const{
     auto foundscore = mem.find(b.GetKey());
     if(foundscore!=mem.end()){
         score = foundscore->second;
